Enable caching on github api calls
Currently github api calls will be cached to the filesystem which may or may
not break things. I had a custom cache into the session but the amount of data
saved was ridiculous.

Also:
The user api usage on the add repo page now gets all users and sorts by type.
Unfortunately repo count is not available from this call, so all users will
be displayed regardless if they have any repositories.